feat(libyml): :art: decoupling `api.rs` via `document.rs`
