+++ tests/test_document.rs
@@ -1,0 +1,132 @@
+#[cfg(test)]
+mod tests {
+    use core::mem::MaybeUninit;
+    use libyml::success::OK;
+    use libyml::yaml_document_delete;
+    use libyml::yaml_document_initialize;
+    use libyml::YamlDocumentT;
+    use libyml::YamlVersionDirectiveT;
+    use std::ptr;
+
+    #[test]
+    fn test_yaml_document_initialize_non_null_document() {
+        unsafe {
+            let mut doc: YamlDocumentT =
+                MaybeUninit::zeroed().assume_init();
+            let result = yaml_document_initialize(
+                &mut doc,
+                ptr::null_mut(),
+                ptr::null_mut(),
+                ptr::null_mut(),
+                false,
+                false,
+            );
+            assert_eq!(
+                result, OK,
+                "Initialization should handle null pointers gracefully"
+            );
+            // Clean up the document to prevent memory leaks.
+            doc.cleanup();
+        }
+    }
+
+    #[test]
+    fn test_yaml_document_initialize() {
+        unsafe {
+            // Allocate a YamlDocumentT on the stack or heap (if more complex setup is needed).
+            let mut doc: YamlDocumentT =
+                MaybeUninit::zeroed().assume_init();
+            let result = yaml_document_initialize(
+                &mut doc,
+                ptr::null_mut(),
+                ptr::null_mut(),
+                ptr::null_mut(),
+                false,
+                false,
+            );
+            // Check if the initialization was successful.
+            assert!(result.ok, "Document initialization should succeed with valid pointer");
+        }
+    }
+    #[test]
+    fn test_yaml_document_delete() {
+        unsafe {
+            // Allocate a YamlDocumentT using MaybeUninit for proper initialization.
+            let mut doc: MaybeUninit<YamlDocumentT> =
+                MaybeUninit::zeroed();
+            let doc_ptr = doc.as_mut_ptr(); // Obtain a mutable pointer to the uninitialized memory.
+
+            // Initialize the document using yaml_document_initialize.
+            let init_result = yaml_document_initialize(
+                doc_ptr,
+                ptr::null_mut(),
+                ptr::null_mut(),
+                ptr::null_mut(),
+                false,
+                false,
+            );
+
+            // Assume yaml_document_initialize returns a boolean or similar, check if initialization was successful.
+            assert!(init_result == OK, "Document initialization should succeed with valid pointer");
+
+            // Since the document is now initialized, you can assume the memory pointed by doc_ptr is properly initialized.
+            let doc = doc_ptr.as_mut().unwrap(); // Convert pointer to reference, which is now safe.
+
+            // Call cleanup method
+            doc.cleanup();
+        }
+    }
+
+    #[test]
+    fn test_yaml_document_initialize_valid() {
+        unsafe {
+            let mut doc: MaybeUninit<YamlDocumentT> =
+                MaybeUninit::uninit();
+            // Create version directive using a hypothetical constructor or mutable allocation
+            let mut version_directive =
+                YamlVersionDirectiveT::new(1, 2); // Assume this constructor exists
+            let mut tag_directives = vec![]; // Example for an empty array of tag directives
+
+            let version_directive_ptr: *mut YamlVersionDirectiveT =
+                &mut version_directive;
+            let result = yaml_document_initialize(
+                doc.as_mut_ptr(),
+                version_directive_ptr,
+                tag_directives.as_mut_ptr(),
+                tag_directives.as_mut_ptr().add(tag_directives.len()),
+                true,
+                false,
+            );
+
+            assert_eq!(
+                result, OK,
+                "Initialization should succeed with valid inputs"
+            );
+            // Assume a cleanup function exists to properly free resources
+            yaml_document_delete(doc.as_mut_ptr());
+        }
+    }
+
+    #[test]
+    fn test_yaml_document_initialize_invalid() {
+        unsafe {
+            let mut doc: MaybeUninit<YamlDocumentT> =
+                MaybeUninit::uninit();
+            let result = yaml_document_initialize(
+                doc.as_mut_ptr(),
+                ptr::null_mut(),
+                ptr::null_mut(),
+                ptr::null_mut(),
+                false,
+                false,
+            );
+
+            assert_eq!(
+                result, OK,
+                "Initialization should handle null pointers gracefully"
+            );
+            // Assume a cleanup function exists to properly free resources
+            yaml_document_delete(doc.as_mut_ptr());
+        }
+    }
+}
